requirements.txt
openai
langchain
faiss-cpu
tiktoken
python-dotenv
duckduckgo-search
newspaper3k
